refactored removed weather and movie functionality from server.js in to seperate modules
