+++ package.json
@@ -21,7 +21,9 @@
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.2",
-    "nodemon": "^2.0.15",
     "react": "^17.0.2"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.15"
   }
 }
